Restore invisible brackets around nicknames
When copying and pasting chat logs from Textual's window, the
brackets help keep the text readable.

They were silently removed during the major overhaul in 796d320.
Because they weren't mentioned at all in the commit message, my
thought is that it might have been accidental—especially since
the CSS was left in.

This is the Textual 6 version of the patch. The version created
for the Textual 5 edition of Equinox was never merged, and now
no longer works.